% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{compute.reliability.measures}
\alias{compute.reliability.measures}
\title{Compute Reliability Measures from Mixed Effects Models}
\usage{
compute.reliability.measures(dat, level = 0.95, B = 500)
}
\arguments{
\item{dat}{The data frame containing the data.}

\item{level}{The confidence level for the confidence intervals.}

\item{B}{The number of bootstrap samples to generate for bootstrapped confidence intervals.}
}
\value{
A list with the intraclass correlation coefficient,
        standard error of measurement, and minimal detectable difference,
        along with their confidence intervals, computed from
        mixed effects models corresponding to ICC(1, 1) and 
        ICC(2, 1) models.
}
\description{
Compute Reliability Measures from Mixed Effects Models
}
